fix: prevent duplicate document ingestion, unsafe string slicing, and stats truncation
Session dedup: begin_session() now checks for completed sessions by file name
+ MD5 before creating new ones, preventing re-upload of already-ingested docs.

Chunk dedup: document_pipeline no longer hardcodes deduplicate=false — uses
SHA256 hash matching to skip duplicate chunks within/between uploads.
Stats truncation: layer-stats, stats, layer-tree, system-status, and MCP all
used ListRequest::default() with limit=10, silently reporting only first 10
memories. Changed to limit=0 (no limit) so commands reflect full bank state.

Layer stats metric: avg_sources/memory now divides by L1+ non-forgotten count
instead of all memories, yielding a meaningful sources-per-abstraction ratio
(1.26 vs 0.10 on real data).

Unsafe slicing: all debug/error string slices replaced with safe_prefix() and
safe_suffix() helpers using char_indices() to avoid UTF-8 boundary panics on
math symbols (², ℓ, etc.).

L1 prompt quality: build_l1_prompt now receives document context (file name,
chunk position, section headers), instructs the LLM to avoid generic meta-
commentary, truncates safely, and encourages concrete summaries.

Search enrichment: pyramid search results now include a neighbors array with
relation type and target ID for each result, surfacing next_chunk/previous_
chunk/part_of/references links.

+++ src/mcp.rs
@@ -993,14 +993,14 @@ impl ServerHandler for MemoryMcpService {
                 }
 
                 let mut total_memories: u64 = 0;
-                let mut by_layer: HashMap<usize, LayerCountStats> = HashMap::new();
+                let mut by_layer: HashMap<i32, LayerCountStats> = HashMap::new();
                 let mut state_counts = json!({
                     "active": 0u64,
                     "forgotten": 0u64,
                     "processing": 0u64,
                     "invalid": 0u64
                 });
-                let mut max_layer: usize = 0;
+                let mut max_layer: i32 = 0;
 
                 // Collect stats from each bank
                 let banks = self.bank_manager.list_banks().await.unwrap_or_default();
@@ -1009,7 +1009,7 @@ impl ServerHandler for MemoryMcpService {
                         && let Ok(memories) = bank.list(&Filters::new(), None).await
                     {
                         for memory in &memories {
-                            let level = memory.metadata.layer.level as usize;
+                            let level = memory.metadata.layer.level;
                             let state = memory.metadata.state.as_str();
 
                             // Update total
